refactor request automations
