+++ app/Modules/Automations/src/Http/Requests/AutomationDestoyRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Modules\Automations\src\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class OrderAutomationDestoyRequest extends FormRequest
+class AutomationDestoyRequest extends FormRequest
 {
     public function authorize():bool
     {
