+++ app/Modules/Automations/src/Http/Requests/AutomationIndexRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Modules\Automations\src\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class OrderAutomationShowRequest extends FormRequest
+class AutomationIndexRequest extends FormRequest
 {
     public function authorize():bool
     {
