+++ app/Modules/Automations/src/Http/Requests/AutomationShowRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Modules\Automations\src\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class OrderAutomationIndexRequest extends FormRequest
+class AutomationShowRequest extends FormRequest
 {
     public function authorize():bool
     {
